Styling template.
Applied a header and footer to the application. Stylized the product cards to be in theme with the rest of the page. The grid is now fully responsive and changes how many product cards are on one row based on viewport width. Fix multiple small styling issues at mobile view ports.