SearchBar
Implemented a working search bar that is case-insensitive and allows partial matches. The search bar code is placed inside a component and exported into the App.jsx. Implemented a no matches found function that displays a "No matches found" message in the grid layout if a users search returns no results.