Filter implementation.
Implemented a filter function. The user is able to filter by in-stock via a checkbox, Category via a drop down menu and price range via two input fields. Applied styling to the side bar and fixed several small styling issues when implementing the new functionality to the side bar. Added more entries to the products array with more category options.