AppHeader, Styling fixes and burger menu.
Implemented a burger menu for iPad and smaller screen sizes. Fixed several styling issues with the grid layout and product cards when searching for different products. Made the footer have a fixed position at the bottom of the page so it is no longer always visible to the user and saves screen space on smaller screens. Moved the header section of the App into it's own component. The app is still fully functional.