Nav styling and sort by
Added a sort by function to the application. A user is now able to sort by price and rating. Each with an option of lowest to highest or highest to lowest. Applied styling to the app navigation. Made two separate search bars. One which is visible on larger viewports and one that is visible on iPad viewports or smaller via the burger menu.